Extract endpoint evidence helpers
